requirements.txt
Pillow==6.1.0
pyTelegramBotAPI==3.6.6
request==2018.11.20
requests==2.22.0
TelegramBotAPI==0.3.2
